+++ template/desktops/xfce.sh
@@ -4,10 +4,10 @@ if [[ -f "${HOME}/.config/monitors.xml" ]]; then
 fi
 
 # Copy over default panel if doesn't exist, otherwise it will prompt the user
-XFCE_CONFIG_ROOT="${HOME}/.config/xfce4/xfconf/xfce-perchannel-xml"
-if [[ ! -e "${XFCE_CONFIG_ROOT}/xfce4-panel.xml" ]]; then
-  mkdir -p "${XFCE_CONFIG_ROOT}"
-  cp "/etc/xdg/xfce4/panel/default.xml" "${XFCE_CONFIG_ROOT}/xfce4-panel.xml"
+PANEL_CONFIG="${HOME}/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-panel.xml"
+if [[ ! -e "${PANEL_CONFIG}" ]]; then
+  mkdir -p "$(dirname "${PANEL_CONFIG}")"
+  cp "/etc/xdg/xfce4/panel/default.xml" "${PANEL_CONFIG}"
 fi
 
 # Disable useless services on autostart
@@ -16,6 +16,20 @@ mkdir -p "${AUTOSTART}"
 for service in "pulseaudio" "rhsm-icon" "spice-vdagent" "tracker-extract" "tracker-miner-apps" "tracker-miner-user-guides" "xfce4-power-manager" "xfce-polkit"; do
   echo -e "[Desktop Entry]\nHidden=true" > "${AUTOSTART}/${service}.desktop"
 done
+
+# Run Xfce4 Terminal as login shell (sets proper TERM)
+TERM_CONFIG="${HOME}/.config/xfce4/terminal/terminalrc"
+if [[ ! -e "${TERM_CONFIG}" ]]; then
+  mkdir -p "$(dirname "${TERM_CONFIG}")"
+  sed 's/^ \{4\}//' > "${TERM_CONFIG}" << EOL
+    [Configuration]
+    CommandLoginShell=TRUE
+EOL
+else
+  sed -i \
+    '/^CommandLoginShell=/{h;s/=.*/=TRUE/};${x;/^$/{s//CommandLoginShell=TRUE/;H};x}' \
+    "${TERM_CONFIG}"
+fi
 
 # Start up xfce desktop (block until user logs out of desktop)
 xfce4-session